Improved README Documentation for BoostNote-Legacy: Installation, Version History, and Contribution Guidelines
Added small description for what the overall product does. Added a detailed installation for the instructions on how to install the product on multiple operating systmes. Provided some details on how to contribute and how developers can get started with contributing in the proper format. Made it much easier to read and more detailed , kept links from original to ensure that if someone needs to do more research into a particular area of the guide, they can.